refactor(modularization): Phase 4 Step 4-S4 (services) — marketplace services move + Q2 check_marketplace_access rename + drop _MarketplaceCollectionOpsAdapter (#25)
Services half of Phase 4 Step 4-S4 (views follow in a paired commit).

**services move**:
- ``git mv`` ``aperag/service/{marketplace,marketplace_collection}_service.py`` to ``aperag/domains/marketplace/service/``.
- G1 boundary pass: rewrote imports to domain paths
  (``aperag.domains.marketplace.db.models`` /
  ``aperag.domains.marketplace.schemas`` /
  ``aperag.domains.knowledge_base.db.models`` cross-domain
  ``Collection`` / ``aperag.domains.knowledge_base.schemas``
  cross-domain ``Document``/``DocumentList``/``DocumentPreview``).

**Q2 public rename (msg=6ab7d211)**:
- ``MarketplaceCollectionService._check_marketplace_access`` →
  ``check_marketplace_access`` (dropped leading underscore). Call
  sites + code comments updated in the same file.
- ``_check_subscription_access`` stays private (internal helper,
  not on the Protocol surface).

**aperag/app.py adapter drop**:
- ``_MarketplaceCollectionOpsAdapter`` removed. The legacy
  marketplace_collection_service now structurally satisfies the
  ``MarketplaceCollectionOps`` Protocol directly (same method
  signature ``check_marketplace_access(user_id, collection_id) ->
  dict``). Wire-up becomes
  ``_kb_set_marketplace_collection_ops(_legacy_marketplace_collection_service)``
  with no adapter layer.

**legacy shims**:
- ``aperag/service/marketplace_service.py`` + ``aperag/service/marketplace_collection_service.py``
  are now pure re-export shims forwarding to the domain path.
- Q2 breaking-change note in the ``marketplace_collection_service``
  shim docstring: callers using the old underscore-prefixed
  ``_check_marketplace_access`` name must migrate.

**Verification**:
- Full unit suite green: 671 passed / 29 skipped.
- ``ruff check`` + ``ruff format --check`` clean.
- Phase 3 KB CI smoke path still passes (KB ``collection_service.create_search``
  now calls the public ``check_marketplace_access`` through the
  wired instance).

4-S4 (views half) follows next: move
``aperag/views/{marketplace,marketplace_collections}.py`` into
``aperag/domains/marketplace/api/`` + rewire ``aperag/app.py``
router includes + OpenAPI regen byte-stable.
